+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to semantic versioning.
 
 ## 0.10.0 (_Unreleased_)
 
+- Add `ReadonlyArray` and `ReadonlyRecord` modules.
 - Add `fromString`, `fromStringWithRadix`, and `floatFromString` functions to the `Number` module.
 
 ## 0.9.0 (2021-03-15)